# src/backend/HomeCloud.Maps.Application/Dto/KomootToursRequest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HomeCloud.Maps.Application.Dto
{
    public class KomootToursRequest
    {
        public string Cookies { get; set; }
    }
}