# src/backend/HomeCloud.Maps.Application/Dto/Tours/PositionDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HomeCloud.Maps.Application.Dto.Tours
{
    public class PositionDto
    {
        public float Latitude { get; set; }

        public float Longitude { get; set; }
    }
}