# src/backend/HomeCloud.Maps.Application/Dto/Tours/TourDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HomeCloud.Maps.Application.Dto.Tours
{
    public class TourDto
    {
        public TourInfoDto TourInfo { get; set; }

        public RouteDto Route { get; set; }
    }
}