# src/frontend/HomeCloud.Maps.Client.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HomeCloud.Maps.Server",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost:44382"
    }
  ],
  "paths": {
    "/api/Routes": {
      "get": {
        "tags": [
          "Routes"
        ],
        "operationId": "GetAllRoutes",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Tour/{tourId}": {
      "get": {
        "tags": [
          "Tour"
        ],
        "operationId": "GetTourInfosById",
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tour": {
      "post": {
        "tags": [
          "Tour"
        ],
        "operationId": "PostTours",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KomootToursRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KomootToursRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KomootToursRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TourInfos": {
      "get": {
        "tags": [
          "TourInfos"
        ],
        "operationId": "GetTourInfosPagination",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "pageIndex",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "tourNameFilter",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourInfoDtoPaginationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourInfoDtoPaginationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourInfoDtoPaginationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/UserSettings": {
      "post": {
        "tags": [
          "UserSettings"
        ],
        "operationId": "PostUserSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "UserSettings"
        ],
        "operationId": "GetUserSettings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "KomootToursRequest": {
        "type": "object",
        "properties": {
          "Cookies": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PositionDto": {
        "type": "object",
        "properties": {
          "Latitude": {
            "type": "number",
            "format": "float"
          },
          "Longitude": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RouteDto": {
        "type": "object",
        "properties": {
          "Positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TourDto": {
        "type": "object",
        "properties": {
          "TourInfo": {
            "$ref": "#/components/schemas/TourInfoDto"
          },
          "Route": {
            "$ref": "#/components/schemas/RouteDto"
          }
        },
        "additionalProperties": false
      },
      "TourInfoDto": {
        "type": "object",
        "properties": {
          "TourId": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Date": {
            "type": "string",
            "format": "date-time"
          },
          "Distance": {
            "type": "number",
            "format": "double"
          },
          "ImageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TourInfoDtoPaginationResult": {
        "type": "object",
        "properties": {
          "Total": {
            "type": "integer",
            "format": "int32"
          },
          "Data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourInfoDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserSettingsDto": {
        "type": "object",
        "properties": {
          "KomootUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}